move TestEsExecutors to package that is accessible to modular code
