+++ server/src/test/java/org/elasticsearch/threadpool/ScalingThreadPoolTests.java
@@ -16,10 +16,10 @@ import org.elasticsearch.common.util.concurrent.EsExecutors;
 import org.elasticsearch.common.util.concurrent.EsRejectedExecutionException;
 import org.elasticsearch.common.util.concurrent.EsRejectedExecutionHandler;
 import org.elasticsearch.common.util.concurrent.EsThreadPoolExecutor;
-import org.elasticsearch.common.util.concurrent.TestEsExecutors;
 import org.elasticsearch.common.util.concurrent.ThreadContext;
 import org.elasticsearch.core.CheckedRunnable;
 import org.elasticsearch.telemetry.metric.MeterRegistry;
+import org.elasticsearch.test.TestEsExecutors;
 import org.hamcrest.Matcher;
 
 import java.util.HashMap;
